feat: uniswap universal router AND refactor: renamed to uniswap from uniswap-pool (negate last commit)
