+++ src/token-aggregators/uniswap/config.ts
@@ -1,6 +1,7 @@
 import {
     AlphaRouter,
     SwapOptionsSwapRouter02,
+    SwapOptionsUniversalRouter,
     SwapType,
 } from '@uniswap/smart-order-router'
 
@@ -10,15 +11,23 @@ import { get_provider } from '@benchmarking-cross-chain-bridges/helper/provider'
 import { CHAIN_ID_MAP } from '@benchmarking-cross-chain-bridges/helper/constants_global';
 import { validate_keys } from '@benchmarking-cross-chain-bridges/helper/inp_validator';
 
-export function create_router(chainId: number): AlphaRouter {
+export function define_alpha_router(chainId: number): AlphaRouter {
     return new AlphaRouter({
         chainId: chainId,
         provider: get_provider(CHAIN_ID_MAP[chainId]),
     });
 }
 
-export function create_swap_options(): SwapOptionsSwapRouter02 {
+export function create_router(router_type: string): SwapOptionsSwapRouter02 | SwapOptionsUniversalRouter {
     const KEY_PUBLIC = validate_keys().public;
+
+    if (router_type === 'universal_router') {
+        return {
+            recipient: KEY_PUBLIC,
+            slippageTolerance: new Percent(50, 10_000),
+            type: SwapType.UNIVERSAL_ROUTER,
+        };
+    }
 
     return {
         recipient: KEY_PUBLIC,
